Handle advisor (& AI) wants as floating point values
See patch #4999

[[originally from svn r27497]]

+++ server/advisors/advtools.h
@@ -13,8 +13,8 @@
 #ifndef FC__ADVTOOLS_H
 #define FC__ADVTOOLS_H
 
-typedef int adv_want;
-#define ADV_WANT_PRINTF "%d"
+typedef float adv_want;
+#define ADV_WANT_PRINTF "%f"
 
 #define MORT 24
